Created a client component to utilize useSearchParams and import the component from search page
