+++ app/search/page.tsx
@@ -1,50 +1,55 @@
+import SearchProduct from "@/components/searchProduct";
+import React, { Suspense } from "react";
+
+export default function SearchPage() {
+  return (
+    <Suspense fallback={<p>載入中…</p>}>
+      <SearchProduct />
+    </Suspense>
+  );
+}
+
 // "use client";
 
 // import { useSearchParams } from "next/navigation";
-import ProductCard from "@/components/products/ProductCard";
+// import ProductCard from "@/components/products/ProductCard";
 // import { useProductStore } from "../stores/useProductStore";
-import { getProducts } from "@/app/utils/actions";
-import type { ProductWithPrice } from "@/app/types/product";
+// import { useEffect } from "react";
 
-export default async function SearchPage({
-  searchParams,
-}: {
-  searchParams: { q?: string };
-}) {
-  // const searchParams = useSearchParams();
-  // const keyword = searchParams.get("q")?.toLowerCase() || "";
-  const keyword = searchParams.q?.toLowerCase() || "";
+// export default function SearchPage() {
+//   const searchParams = useSearchParams();
+//   const keyword = searchParams.get("q")?.toLowerCase() || "";
 
-  // const { allProducts, fetchAllProducts, isLoading } = useProductStore();
-  // server-side fetch
-  const allProducts: ProductWithPrice[] = await getProducts();
+//   const { allProducts, fetchAllProducts, isLoading } = useProductStore();
 
-  // useEffect(() => {
-  //   if (allProducts.length === 0) {
-  //     fetchAllProducts();
-  //   }
-  // }, [fetchAllProducts, allProducts.length]);
+//   useEffect(() => {
+//     if (allProducts.length === 0) {
+//       fetchAllProducts();
+//     }
+//   }, [fetchAllProducts, allProducts.length]);
 
-  // simple filter
-  const filtered = allProducts.filter((product) =>
-    product.title.toLowerCase().includes(keyword)
-  );
+//   // simple filter
+//   const filtered = allProducts.filter((product) =>
+//     product.title.toLowerCase().includes(keyword)
+//   );
 
-  return (
-    <div className="px-4 py-6">
-      <h1 className="text-xl font-bold mb-4">搜尋結果：「{keyword}」</h1>
+//   return (
+//     <div className="px-4 py-6">
+//       <h1 className="text-xl font-bold mb-4">搜尋結果：「{keyword}」</h1>
 
-      {filtered.length > 0 ? (
-        <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-6">
-          {filtered.map((product) => (
-            <ProductCard key={product.id} product={product} />
-          ))}
-        </div>
-      ) : (
-        <p className="text-muted-foreground">
-          {keyword ? "找不到符合的商品" : "請輸入關鍵字進行搜尋"}
-        </p>
-      )}
-    </div>
-  );
-}
+//       {isLoading ? (
+//         <p>載入中…</p>
+//       ) : filtered.length > 0 ? (
+//         <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-6">
+//           {filtered.map((product) => (
+//             <ProductCard key={product.id} product={product} />
+//           ))}
+//         </div>
+//       ) : (
+//         <p className="text-muted-foreground">
+//           {keyword ? "找不到符合的商品" : "請輸入關鍵字進行搜尋"}
+//         </p>
+//       )}
+//     </div>
+//   );
+// }
